aws/requirements.txt
boto3==1.34.77
botocore==1.34.77
requests==2.32.3
beautifulsoup4==4.13.4
soupsieve==2.7
mysql-connector-python==8.3.0